+++ tests/talkpipe/llm/test_embedding.py
@@ -1,4 +1,6 @@
 import pytest
+import json
+from talkpipe.chatterlang import compile
 from talkpipe.llm.embedding import LLMEmbed
 from talkpipe.llm.embedding_adapters import OllamaEmbedderAdapter
 
@@ -16,3 +18,14 @@ def test_create_embedder(requires_ollama):
     assert len(response) == 1
     assert all([isinstance(x, float) for x in response[0]])
     
+def test_lingering_nparray_issue(requires_ollama):
+    f = compile(
+        """
+        | llmEmbed[model="mxbai-embed-large", source="ollama", field="example", append_as="vector"] 
+        | toDict[field_list="example"]
+        """)
+    f = f.asFunction(single_in=True, single_out=False)
+    d = {"example": "Hello"}
+    response = list(f(d))
+    j = json.dumps(response)
+    assert j == '[{"example": "Hello"}]'